изменен тип переменных

+++ WindowsFormsApplication1/Form1.cs
@@ -28,9 +28,9 @@ namespace WindowsFormsApplication1
 
         private void button3_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            float a = Convert.ToSingle(textBox1.Text);
+            float b = Convert.ToSingle(textBox2.Text);
+            float c;
             c = a * b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Умножение";
@@ -53,9 +53,9 @@ namespace WindowsFormsApplication1
 
         private void button4_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            float a = Convert.ToSingle(textBox1.Text);
+            float b = Convert.ToSingle(textBox2.Text);
+            float c;
             c = a / b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Деление";
@@ -63,9 +63,9 @@ namespace WindowsFormsApplication1
 
         private void button2_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            float a = Convert.ToSingle(textBox1.Text);
+            float b = Convert.ToSingle(textBox2.Text);
+            float c;
             c = a - b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Вычитание";
@@ -73,9 +73,9 @@ namespace WindowsFormsApplication1
 
         private void button1_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            float a = Convert.ToSingle(textBox1.Text);
+            float b = Convert.ToSingle(textBox2.Text);
+            float c;
             c = a + b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Сложение";
